pyproject.toml
[project]
name = "verrific"
version = "0.1.1"
description = "Python package for cross-checking and verifying references in scientific papers, designed for use with scienceverse/papercheck"
authors = [
    {name = "Jakub Werner",email = "jakubwerner1@proton.me"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "grobid-client-python (>=0.0.17,<0.0.18)",
]


[poetry.group.dev.dependencies]
jupyter = ">=1.1.1,<2.0.0"
streamlit = ">=1.46.1,<2.0.0"
sphinx = ">=8.2.3,<9.0.0"


[tool.poetry.group.test.dependencies]  
pytest-randomly = "^3.15.0"
pytest = "^8.2.0"
pytest-cov = "^4.1.0"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

